+++ src/backend/joanie/core/models/products.py
@@ -309,12 +309,12 @@ class Product(parler_models.TranslatableModel, BaseModel):
             if courses:
                 offerings = offerings.filter(course__in=courses)
 
-            for relation in offerings.iterator():
+            for offering in offerings.iterator():
                 equivalent_course_runs.append(
                     {
                         **course_run_data,
-                        "resource_link": relation.uri,
-                        "course": relation.course.code,
+                        "resource_link": offering.uri,
+                        "course": offering.course.code,
                     }
                 )
 
@@ -345,7 +345,7 @@ class Product(parler_models.TranslatableModel, BaseModel):
 
             serialized_course_runs.extend(
                 [
-                    course_run.get_serialized(certifying=certifying)
+                    course_run.get_serialized(certifying=certifying, product=product)
                     for course_run in course_runs
                 ]
             )