+++ CHANGELOG.md
@@ -20,6 +20,9 @@ and this project adheres to
 - Add description to order group model
 - Add discounted price to course product relation serializer
 - Add description to order group form in the back office
+- Sync discount str representation
+- Sync course run related discount
+- Trigger sync when an offering rule is saved
 
 ### Changed